requirements.txt
flask==3.0.3
Python-dotenv==1.0.1
httpx==0.27.2